Improved sdp exchange through signaling, fixes

+++ src/app/core/main.py
@@ -6,8 +6,8 @@ from werkzeug.utils import secure_filename
 
 from core import app
 from core.auth import current_user, login_required, load_user
-from core.models import User, Message
-from core.forms import Profile, MessageForm
+from core.models import User, Message, Room
+from core.forms import Profile, MessageForm, RoomForm
 
 
 @app.route('/')
@@ -17,7 +17,7 @@ def home():
 
 
 @app.route('/profile', methods=["GET", "POST"])
-@app.route('/profile/<username>')
+@app.route('/profile/<string:username>')
 @login_required
 def profile(username=None):
     form = Profile()
@@ -33,19 +33,23 @@ def profile(username=None):
         image = form.image.data
         _image_filename, image_file_type = os.path.splitext(image.filename)
 
-        rename_image = secure_filename(f'{str(user.id)}{image_file_type}')
-        for file in os.listdir(app.config['UPLOAD_FOLDER']):
-            if file.startswith(str(user.id)):
-                os.remove(os.path.join(app.config['UPLOAD_FOLDER'], file))
-        image.save(os.path.join(app.config['UPLOAD_FOLDER'], rename_image))
-        
+        if image.filename:
+            rename_image = secure_filename(f'{str(user.id)}{image_file_type}')
+            for file in os.listdir(app.config['UPLOAD_FOLDER']):
+                if file.startswith(str(user.id)):
+                    os.remove(os.path.join(app.config['UPLOAD_FOLDER'], file))
+            image.save(os.path.join(app.config['UPLOAD_FOLDER'], rename_image))
+            
         if form.image.data:
-            user.profile.image = url_for('static', filename=f'img/profile/{rename_image}')
+            user.profile.image = url_for(
+                'static', filename=f'img/profile/{rename_image}'
+            )
         user.profile.fullname = form.fullname.data
         user.profile.description = form.description.data
         user.profile.skills = ", ".join(skills)
         user.profile.about = form.about.data
         user.save()
+        return redirect('profile')
     return render_template('profile.html', form=form, user=user)
 
 
@@ -61,7 +65,7 @@ def notes():
     return render_template('home.html')
 
 @app.route('/inbox', methods=["GET", "POST"])
-@app.route('/inbox/<profile>', methods=["GET", "POST"])
+@app.route('/inbox/<string:profile>', methods=["GET", "POST"])
 @login_required
 def inbox(profile=None):
     form = MessageForm()
@@ -74,8 +78,21 @@ def inbox(profile=None):
                 body=form.body.data
             )
             new_message.save()
+            return redirect('inbox')
     return render_template('inbox.html', form=form, user=current_user)
 
-@app.route('/test/<room>')
-def host(room):
-    return render_template('test.html', room=room, user=current_user)
+@app.route('/room', methods=["GET", "POST"])
+@app.route('/room/<uuid:uuid>', methods=["GET", "POST"])
+@login_required
+def room(uuid=None):
+    form = RoomForm()
+    if form.validate_on_submit():
+        create_room = Room(
+            caller=current_user.username,
+            callee=form.callee.data
+        )
+        create_room.save()
+        uuid = create_room.id
+        return redirect(f'room/{uuid}')
+    room = Room.find_by_id(id=uuid)
+    return render_template('room.html', form=form, room=room, user=current_user)
